[#noissue] Cleanup
