+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/util/RowKeyUtils.java
@@ -69,4 +69,27 @@ public final class RowKeyUtils {
         return rowKey;
     }
 
+    public static byte[] stringLongLongToBytes(final String string, final int maxStringSize, final long value1, final long value2) {
+        return stringLongLongToBytes(0, string, maxStringSize, value1, value2);
+    }
+
+    public static byte[] stringLongLongToBytes(int prefix, final String string, final int maxStringSize, final long value1, final long value2) {
+        if (string == null) {
+            throw new NullPointerException("string");
+        }
+        if (maxStringSize < 0) {
+            throw new StringIndexOutOfBoundsException(maxStringSize);
+        }
+        final byte[] stringBytes = BytesUtils.toBytes(string);
+        if (stringBytes.length > maxStringSize) {
+            throw new StringIndexOutOfBoundsException("string is max " + stringBytes.length + ", string='" + string + "'");
+        }
+        int offset = prefix + maxStringSize;
+        final byte[] buffer = new byte[offset + BytesUtils.LONG_LONG_BYTE_LENGTH];
+        BytesUtils.writeBytes(buffer, prefix, stringBytes);
+        offset = ByteArrayUtils.writeLong(value1, buffer, offset);
+        ByteArrayUtils.writeLong(value2, buffer, offset);
+        return buffer;
+    }
+
 }
